adjust setting routes
